+++ challenges/is-palindrome/solutions/c/function/string.c
@@ -18,7 +18,7 @@ char* string_to_upper(const char* string) {
 char* string_reverse(const char* string) {
   size_t string_length = strlen(string);
   char* result = malloc(sizeof(char) * (string_length + 1));
-  for (size_t index = string_length - 1; index != -1; index--) {
+  for (int index = string_length - 1; index != -1; index--) {
     character_append(result, string[index]);
   }
   return result;
