+++ challenges/print-pyramid/solutions/c/function/solution.c
@@ -12,10 +12,10 @@ int main() {
 
   int step = strcmp(type, "normal") == 0 ? 1 : height;
   while ((strcmp(type, "normal") == 0 && step <= height) || (strcmp(type, "reverse") == 0 && step != 0)) {
-    int numberOfStars = (step * 2) - 1;
-    int totalNumberOfLocations = (height * 2) - 1;
-    int totalNumberOfSpaces = totalNumberOfLocations - numberOfStars;
-    int numberOfSpacesOnEachSide = totalNumberOfSpaces / 2;
+    size_t numberOfStars = (step * 2) - 1;
+    size_t totalNumberOfLocations = (height * 2) - 1;
+    size_t totalNumberOfSpaces = totalNumberOfLocations - numberOfStars;
+    size_t numberOfSpacesOnEachSide = totalNumberOfSpaces / 2;
     character_print(" ", numberOfSpacesOnEachSide);
     character_print("*", numberOfStars);
     character_print(" ", numberOfSpacesOnEachSide);
